[libromdata] Xbox360_XDBF: Explicitly cast std::min() values to `unsigned int`.
MSVC's _byteswap_ulong() returns `unsigned long`, which is technically
the same as `unsigned int` but is considered a different type. This
broke the MSVC build because the two arguments were different types.

+++ src/libromdata/Console/Xbox360_XDBF.cpp
@@ -1731,7 +1731,8 @@ int Xbox360_XDBF::isRomSupported_static(const DetectInfo *info)
 	// SPA will have an XSRC or XSTC entry in the resource table.
 	const unsigned int tblSize = info->header.size - sizeof(*xdbfHeader);
 	const unsigned int entryCount = std::min(
-		be32_to_cpu(xdbfHeader->entry_count), (unsigned int)(tblSize / sizeof(XDBF_Entry)));
+		(unsigned int)be32_to_cpu(xdbfHeader->entry_count),
+		(unsigned int)(tblSize / sizeof(XDBF_Entry)));
 
 	const XDBF_Entry *entry =
 		reinterpret_cast<const XDBF_Entry*>(&info->header.pData[sizeof(*xdbfHeader)]);
